requirements.txt
Flask
cryptography
flask_jwt_extended
prometheus_client
python-dotenv
